fixed graph_types and graph_utils tests
